♻️ refactor : 폴더 재구조화 및 명칭 수정
