+++ src/main/java/com/zipte/notifications/server/application/service/task/PropertyNotificationService.java
@@ -1,12 +1,11 @@
-package com.zipte.notifications.server.application.service;
+package com.zipte.notifications.server.application.service.task;
 
 import com.zipte.notifications.core.utils.NotificationIdGenerator;
 import com.zipte.notifications.server.adapter.out.mongo.base.NotificationType;
 import com.zipte.notifications.server.application.port.in.task.AddPropertyNotificationTask;
 import com.zipte.notifications.server.application.port.in.task.RemovePropertyNotificationTask;
-import com.zipte.notifications.server.application.port.out.DeletePropertyPort;
-import com.zipte.notifications.server.application.port.out.LoadPropertyPort;
-import com.zipte.notifications.server.application.port.out.SavePropertyPort;
+import com.zipte.notifications.server.application.port.out.task.DeletePropertyPort;
+import com.zipte.notifications.server.application.port.out.task.SavePropertyPort;
 import com.zipte.notifications.server.adapter.in.consumer.dto.PropertyEvent;
 import com.zipte.notifications.server.domain.PropertyNotification;
 import lombok.RequiredArgsConstructor;
@@ -26,7 +25,6 @@ import java.time.temporal.ChronoUnit;
 public class PropertyNotificationService implements AddPropertyNotificationTask, RemovePropertyNotificationTask {
 
     private final SavePropertyPort savePort;
-    private final LoadPropertyPort loadPort;
     private final DeletePropertyPort deletePort;
 
     @Override
